+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.2.0 (upcoming)
+
+- Added acf helper class
+- Added repeater helper function
+
 ## 0.1.1 (released 2017-05-09)
 
 - Fix bug if acf isn't available